pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "oa-sdk-py"
version = "0.1.0a0"
description = "Python SDK for The Open Anonymity Project"
readme = "README.md"
requires-python = ">=3.10"
authors = [
  { name = "The Open Anonymity Project" }
]
license = { text = "MIT" }
dependencies = [
  "httpx>=0.27.0"
]

[project.optional-dependencies]
blind-signatures = [
  "privacypass-py"
]
gemini-live = [
  "google-genai>=1.0.0"
]
dev = [
  "pytest>=8.0.0",
  "pytest-asyncio>=0.23.0",
  "ruff>=0.6.0",
  "mypy>=1.10.0"
]

[project.scripts]
oa-sdk-py = "oa_sdk.cli:main"
oa-sdk = "oa_sdk.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
markers = [
  "e2e_live: real network end-to-end tests that consume real tickets",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
strict = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_return_any = true
